chiselc: Move extract_filter() to transforms/utils.rs
